Optimize report output process
