+++ pkg/transform/api_transform.go
@@ -8,6 +8,7 @@ import (
 	"github.com/fusor/cpma/pkg/decode"
 	"github.com/fusor/cpma/pkg/env"
 	"github.com/fusor/cpma/pkg/io"
+	"github.com/fusor/cpma/pkg/transform/reportoutput"
 	"github.com/sirupsen/logrus"
 )
 
@@ -33,17 +34,13 @@ func (e APIExtraction) Transform() ([]Output, error) {
 	outputs := []Output{}
 	if env.Config().GetBool("Reporting") {
 		logrus.Info("APITransform::Transform:Reports")
-		reports, err := e.buildReportOutput()
-		if err != nil {
-			return nil, err
-		}
-		outputs = append(outputs, reports)
+		e.buildReportOutput()
 	}
 	return outputs, nil
 }
 
-func (e APIExtraction) buildReportOutput() (Output, error) {
-	componentReport := ComponentReport{
+func (e APIExtraction) buildReportOutput() {
+	componentReport := reportoutput.ComponentReport{
 		Component: APIComponentName,
 	}
 
@@ -56,7 +53,7 @@ func (e APIExtraction) buildReportOutput() (Output, error) {
 	}
 
 	componentReport.Reports = append(componentReport.Reports,
-		Report{
+		reportoutput.Report{
 			Name:       "API",
 			Kind:       "Port",
 			Supported:  false,
@@ -64,11 +61,7 @@ func (e APIExtraction) buildReportOutput() (Output, error) {
 			Comment:    fmt.Sprintf("The API Port for Openshift 4 is 6443 and is non-configurable. Your OCP 3 cluster is currently configured to use port %v", port),
 		})
 
-	reportOutput := ReportOutput{
-		ComponentReports: []ComponentReport{componentReport},
-	}
-
-	return reportOutput, nil
+	finalReportOutput.report.ComponentReports = append(finalReportOutput.report.ComponentReports, componentReport)
 }
 
 // Extract collects API configuration from an OCP3 cluster
